Refactor and enhance Anyclick DevTools package
- Introduced the @ewjdev/anyclick-devtools package for advanced inspection capabilities, including a full-featured InspectDialog and modification tracking.
- Updated the AnyclickProviderWrapper to integrate the new devtools package.
- Enhanced the overall structure and documentation of the Anyclick ecosystem, including updates to existing components and styles.
- Added new utilities for highlighting and IDE integration, improving user experience during feedback capture.
