+++ .changeset/seven-nails-sleep.md
@@ -1,0 +1,6 @@
+---
+"@ewjdev/anyclick-react": minor
+"@ewjdev/anyclick-devtools": patch
+---
+
+move inspect into seperate package for devtools
