+++ .prettierrc
@@ -1,3 +1,12 @@
 {
-  "arrowParens": "always"
+  "arrowParens": "always",
+  "printWidth": 80,
+  "semi": true,
+  "singleQuote": false,
+  "tabWidth": 2,
+  "trailingComma": "all",
+  "plugins": ["@trivago/prettier-plugin-sort-imports"],
+  "importOrder": ["^react$", "^react-dom$", "^@?\\w", "^@ewjdev/", "^[./]"],
+  "importOrderSeparation": false,
+  "importOrderSortSpecifiers": true
 }
