+++ apps/web/package.json
@@ -13,6 +13,7 @@
     "@headlessui/react": "^2.2.8",
     "@heroicons/react": "^2.1.3",
     "@ewjdev/anyclick-core": "*",
+    "@ewjdev/anyclick-devtools": "*",
     "@ewjdev/anyclick-cursor": "*",
     "@ewjdev/anyclick-adapters": "*",
     "@ewjdev/anyclick-github": "*",
