+++ packages/anyclick-devtools/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "@ewjdev/anyclick-react",
-  "version": "1.4.1",
+  "name": "@ewjdev/anyclick-devtools",
+  "version": "1.4.0",
   "private": false,
-  "description": "React provider and context menu UI for UI feedback capture",
+  "description": "Advanced anyclick inspector and devtools UI (full InspectDialog)",
   "keywords": [
     "feedback",
-    "ui",
-    "context-menu",
+    "inspector",
+    "devtools",
     "react",
     "typescript"
   ],
@@ -19,9 +19,9 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/ewjdev/anyclick.git",
-    "directory": "packages/anyclick-react"
+    "directory": "packages/anyclick-devtools"
   },
-  "homepage": "https://anyclick.ewj.dev/docs/react",
+  "homepage": "https://anyclick.ewj.dev/docs/devtools",
   "bugs": {
     "url": "https://github.com/ewjdev/anyclick/issues"
   },
@@ -31,11 +31,6 @@
   "types": "dist/index.d.ts",
   "exports": {
     ".": {
-      "types": "./dist/index.d.ts",
-      "require": "./dist/index.js",
-      "import": "./dist/index.mjs"
-    },
-    "./presets": {
       "types": "./dist/index.d.ts",
       "require": "./dist/index.js",
       "import": "./dist/index.mjs"
@@ -51,7 +46,7 @@
   },
   "dependencies": {
     "@ewjdev/anyclick-core": "^1.4.0",
-    "@ewjdev/anyclick-pointer": "^1.4.0",
+    "@ewjdev/anyclick-react": "^1.4.0",
     "lucide-react": "^0.475.0",
     "zustand": "^5.0.2"
   },
