+++ packages/anyclick-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ewjdev/anyclick-react
 
+## 1.4.1
+
+### Patch Changes
+
+- Added documentation and cleanup
+
 ## 1.4.0
 
 ### Patch Changes
